requirements.txt
tensorflow
numpy
opencv-python
matplotlib
pillow
requests
